更新版本

+++ gdal/frmts/grib/degrib18/degrib/memorydatasource.cpp
@@ -25,7 +25,7 @@ size_t MemoryDataSource::DataSourceFread(void* lpBuf, size_t size, size_t count)
 		eof = false; // feof also "resets" after a good read
 
 	memcpy(lpBuf, memoryBlock + seekPos, size * count);
-	seekPos += size * count;
+	seekPos += static_cast<long>(size * count);
 
 	return count;
