+++ gdal/frmts/grib/degrib18/g2clib-1.0.4/seekgb.c
@@ -25,7 +25,7 @@ void seekgb(FILE *lugb,g2int iseek,g2int mseek,g2int *lskip,g2int *lgrib)
 //     iseek      - number of bytes in the file to skip before search
 //     mseek      - number of bytes to search at a time
 //   OUTPUT ARGUMENTS:
-//     lskip      - number of bytes to skip from the beggining of the file
+//     lskip      - number of bytes to skip from the beginning of the file
 //                  to where the GRIB message starts
 //     lgrib      - number of bytes in message (set to 0, if no message found)
 //
@@ -52,7 +52,7 @@ void seekgb(FILE *lugb,g2int iseek,g2int mseek,g2int *lskip,g2int *lgrib)
 //  READ PARTIAL SECTION
 
         /* ret= */ fseek(lugb,ipos,SEEK_SET);
-        nread=fread(cbuf,sizeof(unsigned char),mseek,lugb);
+        nread=(int)fread(cbuf,sizeof(unsigned char),mseek,lugb);
         lim=nread-8;
 
 //  LOOK FOR 'GRIB...' IN PARTIAL SECTION
@@ -65,7 +65,7 @@ void seekgb(FILE *lugb,g2int iseek,g2int mseek,g2int *lskip,g2int *lgrib)
             if (vers == 1) gbit(cbuf,&lengrib,(k+4)*8,3*8);
             if (vers == 2) gbit(cbuf,&lengrib,(k+12)*8,4*8);
             /* ret= */ fseek(lugb,ipos+k+lengrib-4,SEEK_SET);
-            k4=fread(&end,sizeof(g2int),1,lugb);
+            k4=(int)fread(&end,sizeof(g2int),1,lugb);
             if (k4 == 1 && end == 926365495) {      //GRIB message found
                 *lskip=ipos+k;
                 *lgrib=lengrib;
